add publication tabs
